refactor(modules): consolidate the delegate subsystem into src/modules/delegates/
Finish the delegates module: move 24 server/delegate_*.c core files (+ their
headers) into modules/delegates/, which previously held only the orchestration
seam (gw_orch_delegates). Mirrors the economizer #1485 finish.

modules/delegates/ <- server/delegate_{backend*,checkout,credential_retry,
credentials,depth,driver,economics,ensemble,ensemble_review,ephemeral_ws,
launch,openai,patch_coordinator,prompt,prompt_stdin,role,routing,run_phases,
sandbox_image,source_authority,xml_fallback}.c + their headers.

Pure relocate: git mv + build-path updates, no logic change. Moved files use
bare includes (resolve via the existing -Imodules/delegates); stay-behind
prefixed includes of moved headers rewritten to bare (roadmap_auto.c;
tests/test_anthropic_http*.c which used ../headers/delegate_driver.h). Build
refs repathed on both build systems, all seds anchored on `server/delegate_`
so the separate server/server_delegate_* subsystem is untouched.

DEFERRED delegate_plan.c/.h: kept at server/headers. Relocating it surfaced a
pre-existing latent uninitialized-local-variable bug in delegate_plan.c's
owned-path parse (feeds implementation_packet_count) that only "passes by
memory-layout luck" — the __FILE__ length change flipped the garbage value and
unit-test-delegate-plan failed deterministically. Confirmed uninit
(-ftrivial-auto-var-init=zero fixes it; ASAN/UBSAN clean). It needs a dedicated
fix (MSAN trace) before it can be moved; out of scope for a mechanical relocate.

Stays behind (different concern): server/server_delegate_* (monitor/ondemand/
status), server/{gateway,wfe_live,wfe_delegate_resolve}_delegate*, the
cmd_agent_delegate* CLI commands, db1/delegate_learning, and the test stubs.

Server links; unit-test-delegate-* (incl. delegate-plan) pass; make
build-integrity + make lint green.
