+++ src/modules/delegates/delegate_credential_retry.h
