+++ src/modules/delegates/delegate_credentials.c
