+++ src/modules/delegates/delegate_credentials.h
