+++ src/modules/delegates/delegate_ensemble.c
