+++ src/modules/delegates/delegate_ensemble.h
