+++ src/modules/delegates/delegate_ensemble_internal.h
