+++ src/modules/delegates/delegate_launch.c
