+++ src/modules/delegates/delegate_openai.c
