+++ src/modules/delegates/delegate_prompt.c
