+++ src/modules/delegates/delegate_prompt_stdin.c
