+++ src/modules/delegates/delegate_sandbox_image.h
