+++ src/modules/delegates/delegate_source_authority.h
