+++ src/modules/delegates/delegate_xml_fallback.c
